+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fix
+- Fixed missing command to retrieve PV power forecast when using list of values.
+
 ## [0.2.3] - 2022-03-29
 ### Added
 - Improved support for the new add-on and direct communication via the supervisor.