+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.2.4] - 2022-04-12
 ### Fix
 - Fixed missing command to retrieve PV power forecast when using list of values.
 - Updated handling of freq definition to a pandas index.
@@ -88,3 +88,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.2.1]: https://github.com/davidusb-geek/emhass/releases/tag/v0.2.1
 [0.2.2]: https://github.com/davidusb-geek/emhass/releases/tag/v0.2.2
 [0.2.3]: https://github.com/davidusb-geek/emhass/releases/tag/v0.2.3
+[0.2.4]: https://github.com/davidusb-geek/emhass/releases/tag/v0.2.4